+++ packages/number-treemap/src/TreeMap.js
@@ -9,74 +9,11 @@ class TreeMap {
         this.size = 0;
     }
 
-    get(key) {
-        let currentNode = this.root;
-        while (currentNode != null) {
-            if (key < currentNode.key) {
-                currentNode = currentNode.left;
-            } else if (key > currentNode.key) {
-                currentNode = currentNode.right;
-            } else {
-                return currentNode.value;
-            }
-        }
-        return null;
-    }
 
-    getClosestEntry(key) {
-        let currentNode = this.root;
-        let bestMatch = currentNode;
-        let bestDistance = Number.POSITIVE_INFINITY;
-        while (currentNode != null) {
-            let distance = Math.abs(currentNode.key - key);
-            if (key < currentNode.key) {
-                if (distance < bestDistance) {
-                    bestMatch = currentNode;
-                    bestDistance = distance;
-                }
-                currentNode = currentNode.left;
-            } else if (key > currentNode.key) {
-                if (distance < bestDistance) {
-                    bestMatch = currentNode;
-                    bestDistance = distance;
-                }
-                currentNode = currentNode.right;
-            } else {
-                return currentNode;
-            }
-        }
-        return bestMatch;
-
-    }
-
-    set(key, value) {
-        if (this.root === null) {
-            this.root = new Node(key, value);
-            this.size = 1;
-            return;
-        }
-        let currentNode = this.root;
-        let parent;
-        do {
-            parent = currentNode;
-            if (key < currentNode.key) {
-                currentNode = currentNode.left;
-            } else if (key > currentNode.key) {
-                currentNode = currentNode.right;
-            } else {
-                currentNode.value = value;
-                return;
-            }
-        } while (currentNode !== null);
-        let node = new Node(key, value);
-        if (key < parent.key) {
-            parent.left = node;
-        } else {
-            parent.right = node;
-        }
-        this.size++;
-    }
 }
 
+TreeMap.prototype.get = require('./get.js');
+TreeMap.prototype.set = require('./set.js');
+TreeMap.prototype.getClosestEntry = require('./getClosestEntry.js');
 
 module.exports = TreeMap;
